Changed User trying to access database
This means you probably shouldn't have to change password in application.proprties to your localhost password? Although you may need to re-run the dump file to run the sql that creates the user on local host and gives it access to the database. It currently gives USER adminuser@localhost all privileges on mydb.* (meaning all tables in mydb database).